Separate TextField from MoreConfig
