+++ app/javascript/src/components/common-prop-types/index.js
@@ -5,3 +5,4 @@ export * from './tag-prop-types';
 export * from './article-prop-types';
 export * from './cover-image-prop-types';
 export * from './image-management-prop-types';
+export * from './text-field-prop-types';
