+++ app/javascript/src/components/common-prop-types/text-field-prop-types.js
@@ -1,0 +1,8 @@
+import PropTypes from 'prop-types';
+
+export const textFieldPropTypes = PropTypes.shape({
+  label: PropTypes.string.isRequired,
+  id: PropTypes.string.isRequired,
+  value: PropTypes.string.isRequired,
+  onKeyUp: PropTypes.func.isRequired,
+});
